- Create `PokemonRepository` interface in `core-domain` module. - Move `PokemonRepository` implementation to `core-domain` module. - Remove `toDomain` mapper in `GetPokemonListUseCase`. - Remove `toDomain` mapper in `GetPokemonDetailsUseCase`. - Update `GetPokemonListUseCase` and `GetPokemonDetailsUseCase` to receive a Result instead a DTO.
