+++ core-domain/src/main/java/br/ftdev/core/domain/repository/PokemonRepository.kt
@@ -1,0 +1,10 @@
+package br.ftdev.core.domain.repository
+
+import br.ftdev.core.domain.model.Pokemon
+import br.ftdev.core.domain.model.PokemonDetails
+
+interface PokemonRepository {
+    suspend fun getPokemonList(limit: Int, offset: Int): Result<List<Pokemon>>
+    suspend fun getPokemonDetails(nameOrId: String): Result<PokemonDetails>
+    suspend fun refreshPokemonList(): Result<Unit>
+}
