+++ core-domain/src/main/java/br/ftdev/core/domain/usecase/GetPokemonDetailsUseCase.kt
@@ -1,8 +1,7 @@
 package br.ftdev.core.domain.usecase
 
-import br.ftdev.core.data.repository.PokemonRepository
-import br.ftdev.core.domain.mapper.toDomain
 import br.ftdev.core.domain.model.PokemonDetails
+import br.ftdev.core.domain.repository.PokemonRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 
@@ -11,8 +10,8 @@ class GetPokemonDetailsUseCase(
 ) {
     suspend operator fun invoke(nameOrId: String): Flow<Result<PokemonDetails>> = flow {
         pokemonRepository.getPokemonDetails(nameOrId)
-            .onSuccess { detailsDto ->
-                emit(Result.success(detailsDto.toDomain()))
+            .onSuccess { result ->
+                emit(Result.success(result))
             }
             .onFailure { exception ->
                 emit(Result.failure(exception))
